+++ src/controllers/admin/adminDashboard.controller.js
@@ -1,0 +1,11 @@
+const loadDashboard = async (req, res) => {
+  res.render("admin/dashboard", {
+    layout: "layouts/adminLayout",
+    activePage: "dashboard",
+    admin:req.admin,
+  });
+};
+
+module.exports = {
+    loadDashboard,
+}
